+++ README.md
@@ -83,5 +83,5 @@
 
 ### 체크 리스트
 * [x] Domain 패키지로 이동
-* [ ] View 패키지로 이동
+* [x] View 패키지로 이동
 * [ ] View -> Domain 으로만 의존하도록 변경
