+++ src/test/kotlin/car/view/InputViewTest.kt
@@ -1,4 +1,4 @@
-package car.racing
+package car.view
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FreeSpec
